[WEB-4514]chore: moved EIssueLayoutTypes enum to types #7416
