+++ apps/web/core/components/dropdowns/layout.tsx
@@ -2,13 +2,15 @@ import { useCallback, useMemo } from "react";
 import { observer } from "mobx-react";
 import { Check } from "lucide-react";
 // plane constants
-import { EIssueLayoutTypes, ISSUE_LAYOUT_MAP } from "@plane/constants";
+import { ISSUE_LAYOUT_MAP } from "@plane/constants";
 // plane i18n
 import { useTranslation } from "@plane/i18n";
 // plane ui
 import { Dropdown } from "@plane/ui";
 // plane utils
 import { cn } from "@plane/utils";
+// types
+import { EIssueLayoutTypes } from "@plane/types";
 // components
 import { IssueLayoutIcon } from "@/components/issues";
