+++ apps/web/core/components/issues/issue-layouts/gantt/base-gantt-root.tsx
@@ -2,9 +2,9 @@ import React, { useCallback, useEffect } from "react";
 import { observer } from "mobx-react";
 import { useParams } from "next/navigation";
 // plane constants
-import { ALL_ISSUES, EIssueLayoutTypes, EUserPermissions, EUserPermissionsLevel } from "@plane/constants";
+import { ALL_ISSUES, EUserPermissions, EUserPermissionsLevel } from "@plane/constants";
 import { useTranslation } from "@plane/i18n";
-import { EIssuesStoreType, IBlockUpdateData, TIssue } from "@plane/types";
+import { EIssuesStoreType, IBlockUpdateData, TIssue, EIssueLayoutTypes } from "@plane/types";
 import { setToast, TOAST_TYPE } from "@plane/ui";
 // hooks
 import { renderFormattedPayloadDate } from "@plane/utils";
