+++ apps/web/core/components/issues/issue-layouts/list/list-group.tsx
@@ -5,7 +5,7 @@ import { combine } from "@atlaskit/pragmatic-drag-and-drop/combine";
 import { dropTargetForElements } from "@atlaskit/pragmatic-drag-and-drop/element/adapter";
 import { observer } from "mobx-react";
 // plane constants
-import { EIssueLayoutTypes, DRAG_ALLOWED_GROUPS } from "@plane/constants";
+import {  DRAG_ALLOWED_GROUPS } from "@plane/constants";
 // plane i18n
 import { useTranslation } from "@plane/i18n";
 // plane ui
@@ -17,6 +17,7 @@ import {
   TIssue,
   IIssueDisplayProperties,
   TIssueKanbanFilters,
+  EIssueLayoutTypes,
 } from "@plane/types";
 import { Row, setToast, TOAST_TYPE } from "@plane/ui";
 // plane utils
