+++ apps/web/core/components/issues/issue-layouts/quick-add/root.tsx
@@ -6,10 +6,10 @@ import { useParams } from "next/navigation";
 import { useForm, UseFormRegister } from "react-hook-form";
 import { PlusIcon } from "lucide-react";
 // plane constants
-import { EIssueLayoutTypes, WORK_ITEM_TRACKER_EVENTS } from "@plane/constants";
+import {  WORK_ITEM_TRACKER_EVENTS } from "@plane/constants";
 // i18n
 import { useTranslation } from "@plane/i18n";
-import { IProject, TIssue } from "@plane/types";
+import { IProject, TIssue, EIssueLayoutTypes } from "@plane/types";
 // ui
 import { setPromiseToast } from "@plane/ui";
 import { cn, createIssuePayload } from "@plane/utils";
