+++ apps/web/core/components/issues/issue-layouts/roots/module-layout-root.tsx
@@ -3,8 +3,7 @@ import { observer } from "mobx-react";
 import { useParams } from "next/navigation";
 import useSWR from "swr";
 // plane constants
-import { EIssueLayoutTypes } from "@plane/constants";
-import { EIssuesStoreType } from "@plane/types";
+import { EIssuesStoreType, EIssueLayoutTypes } from "@plane/types";
 // components
 import { Row, ERowVariant } from "@plane/ui";
 import { LogoSpinner } from "@/components/common";
