+++ apps/web/core/components/issues/issue-layouts/roots/project-layout-root.tsx
@@ -5,8 +5,7 @@ import { observer } from "mobx-react";
 import { useParams } from "next/navigation";
 import useSWR from "swr";
 // plane constants
-import { EIssueLayoutTypes } from "@plane/constants";
-import { EIssuesStoreType } from "@plane/types";
+import { EIssuesStoreType, EIssueLayoutTypes } from "@plane/types";
 // components
 import { Spinner } from "@plane/ui";
 import { LogoSpinner } from "@/components/common";
