+++ apps/web/core/components/issues/issue-layouts/spreadsheet/roots/workspace-root.tsx
@@ -1,14 +1,8 @@
 import React, { useCallback } from "react";
 import { observer } from "mobx-react";
 // plane constants
-import {
-  ALL_ISSUES,
-  EIssueLayoutTypes,
-  EIssueFilterType,
-  EUserPermissions,
-  EUserPermissionsLevel,
-} from "@plane/constants";
-import { IIssueDisplayFilterOptions, EIssuesStoreType } from "@plane/types";
+import { ALL_ISSUES, EIssueFilterType, EUserPermissions, EUserPermissionsLevel } from "@plane/constants";
+import { IIssueDisplayFilterOptions, EIssuesStoreType, EIssueLayoutTypes } from "@plane/types";
 // components
 import { SpreadsheetView } from "@/components/issues/issue-layouts";
 import { AllIssueQuickActions } from "@/components/issues/issue-layouts/quick-action-dropdowns";
