+++ packages/types/src/issues/issue.ts
@@ -5,6 +5,14 @@ import { TIssueReaction, IIssuePublicReaction, IPublicVote } from "./issue_react
 import { TIssueRelationTypes } from "./issue_relation";
 import { TIssuePublicComment } from "./activity/issue_comment";
 
+export enum EIssueLayoutTypes {
+  LIST = "list",
+  KANBAN = "kanban",
+  CALENDAR = "calendar",
+  GANTT = "gantt_chart",
+  SPREADSHEET = "spreadsheet",
+}
+
 export enum EIssueServiceType {
   ISSUES = "issues",
   EPICS = "epics",
